refactoring: Integrate request DTOs for report and like requests in post and comment
